backend/pyproject.toml
[project]
name = "lmutry-backend"
version = "0.1.0"
description = "Backend para análisis de telemetría de Le Mans Ultimate"
requires-python = ">=3.10"
dependencies = [
    "duckdb>=0.10.0",
    "pandas>=2.0.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-multipart>=0.0.6",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.2.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "analysis*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
